pyproject.toml
[project]
name = "claude-usage-plz"
version = "0.1.0"
description = "Check your Claude Max/Pro usage limits from the command line"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
dependencies = [
    "pexpect",
    "pyte",
]

[project.scripts]
claude-usage = "claude_usage:main"

[project.urls]
Homepage = "https://github.com/hyang0129/claude-usage-plz"
Issues = "https://github.com/hyang0129/claude-usage-plz/issues"

[tool.pytest.ini_options]
markers = ["integration: live tests requiring authenticated claude"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
